+++ R/class_network.R
@@ -128,10 +128,12 @@ network_class <- R6::R6Class(
       }
       self$update_targets()
       vertices <- rbind(self$vertices_targets, self$vertices_imports)
-      vertices <- vertices[!duplicated(vertices$name), ]
+      vertices <- vertices[!duplicated(vertices$name),, drop = FALSE] # nolint
+      vertices <- vertices[order(vertices$name),, drop = FALSE] # nolint
       edges <- rbind(self$edges_imports, self$edges_targets)
       edges <- edges[edges$from %in% vertices$name,, drop = FALSE] # nolint
       edges <- edges[edges$to %in% vertices$name,, drop = FALSE] # nolint
+      edges <- edges[order(edges$from),, drop = FALSE] # nolint
       self$vertices <- vertices
       self$edges <- edges
       self$shortcut_vertices()