+++ R/class_visual.R
@@ -100,6 +100,16 @@ visual_class <- R6::R6Class(
       vertices <- self$network$vertices
       vertices$color <- self$produce_fills(vertices$status)
       self$network$vertices <- vertices
+      colors <- vertices[, c("name", "color")]
+      colors$to <- colors$name
+      colors$name <- NULL
+      self$network$edges <- merge(
+        x = self$network$edges,
+        y = colors,
+        by = "to",
+        all.x = TRUE,
+        all.y = FALSE
+      )[, c("from", "to", "color")]
     },
     update_legend = function() {
       self$legend <- self$produce_legend()